+++ tests/test_smoke.py
@@ -11,7 +11,8 @@ import json
 import pytest
 
 TOOLS = {
-    "mem_remember", "mem_fact", "mem_link", "mem_graph_query", "mem_recall", "mem_expand",
+    "mem_remember", "mem_fact", "mem_annotate", "mem_link", "mem_graph_query",
+    "mem_recall", "mem_expand",
     "mem_update", "mem_compact", "mem_assemble", "mem_forget", "mem_reindex",
     "mem_recent", "mem_evidence", "mem_batch", "mem_get", "mem_inspect",
     "mem_load_session", "mem_status", "mem_doctor",